+++ src/Ampersand/Commands/Proto.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE NoImplicitPrelude #-}
+﻿{-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE DeriveDataTypeable #-}
 {-# LANGUAGE FlexibleContexts #-}
@@ -14,26 +14,28 @@ module Ampersand.Commands.Proto
 import           Ampersand.Basics
 import           Ampersand.FSpec
 import           Ampersand.Misc.HasClasses
-import           Ampersand.Output.FSpec2SQL
-import           Ampersand.Output.ToJSON.ToJson
-import           Ampersand.Prototype.GenFrontend (doGenFrontend)
+import           Ampersand.Prototype.GenFrontend (doGenFrontend, doGenBackend, copyCustomizations)
 import           Ampersand.Types.Config
 import qualified RIO.Text as T
 import           System.Directory
 -- | Builds a prototype of the current project.
 --
-proto :: (Show env, HasRunner env, HasRunComposer env, HasDirCustomizations env, HasZwolleVersion env, HasProtoOpts env, HasDirPrototype env) 
+proto :: (Show env, HasRunner env, HasDirCustomizations env, HasZwolleVersion env, HasProtoOpts env, HasDirPrototype env, HasGenerateFrontend env, HasGenerateBackend env) 
        => FSpec -> RIO env ()
 proto fSpec = do
     env <- ask
     let dirPrototype = getDirPrototype env
     logDebug "Generating prototype..."
     liftIO $ createDirectoryIfMissing True dirPrototype
-    doGenFrontend fSpec
-    generateDatabaseFile fSpec
-    let dir = getGenericsDir env
-    generateAllJSONfiles dir fSpec
+    generateFrontend <- view generateFrontendL
+    generateBackend <- view generateBackendL
+    if generateFrontend 
+     then do doGenFrontend fSpec
+     else do logDebug "  Skipping generating frontend files"
+    if generateBackend
+      then do doGenBackend fSpec
+      else do logDebug "  Skipping generating backend files"
+    copyCustomizations
     dirPrototypeA <- liftIO $ makeAbsolute dirPrototype
     logInfo $ "Prototype files have been written to " <> display (T.pack dirPrototypeA)
 
-