+++ src/Ampersand/Misc/HasClasses.hs
@@ -1,10 +1,11 @@
-{-# LANGUAGE FlexibleInstances #-}
+﻿{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE UndecidableInstances #-}
 module Ampersand.Misc.HasClasses
 
 where
 import Ampersand.Basics
+import Ampersand.Misc.Defaults (defaultDirPrototype)
 import RIO.FilePath
 
 class HasFSpecGenOpts a where
@@ -57,7 +58,7 @@ class (HasRootFile a) => HasDirPrototype a where
   getDirPrototype :: a -> FilePath
   getDirPrototype x =
     (case view dirPrototypeL x of
-       Nothing -> fromMaybe "" (view (rootFileL) x) <> ".proto"
+       Nothing -> defaultDirPrototype
        Just nm -> nm )
 instance HasDirPrototype ProtoOpts where
   dirPrototypeL = lens xdirPrototype (\x y -> x { xdirPrototype = y })
@@ -66,6 +67,16 @@ class HasAllowInvariantViolations a where
   allowInvariantViolationsL :: Lens' a Bool
 instance (HasFSpecGenOpts a) => HasAllowInvariantViolations a where
   allowInvariantViolationsL = fSpecGenOptsL . (lens xallowInvariantViolations (\x y -> x { xallowInvariantViolations = y }))
+class HasGenerateFrontend a where
+  generateFrontendL :: Lens' a Bool
+instance HasGenerateFrontend ProtoOpts where
+  generateFrontendL = lens xgenerateFrontend (\x y -> x { xgenerateFrontend = y })
+
+class HasGenerateBackend a where
+  generateBackendL :: Lens' a Bool
+instance HasGenerateBackend ProtoOpts where
+  generateBackendL = lens xgenerateBackend (\x y -> x { xgenerateBackend = y })
+
 class HasRootFile a where
   rootFileL :: Lens' a (Maybe FilePath)
   baseName :: a -> FilePath
@@ -115,14 +126,8 @@ instance HasOutputLanguage DocOpts where
 instance HasOutputLanguage UmlOpts where
   languageL = lens x4OutputLanguage (\x y -> x { x4OutputLanguage = y })
 
-class HasRunComposer a where
-  skipComposerL :: Lens' a Bool -- if True, runs Composer (php package manager) when generating prototype. Requires PHP and Composer on the machine. Added as switch to disable when building with Docker.
-instance HasRunComposer ProtoOpts where
-  skipComposerL = lens xskipComposer (\x y -> x { xskipComposer = y })
-
-
 class HasDirCustomizations a where
-  dirCustomizationsL :: Lens' a [FilePath] -- the directories that are copied after generating the prototype
+  dirCustomizationsL :: Lens' a (Maybe [FilePath]) -- the directories that are copied after generating the prototype
 instance HasDirCustomizations ProtoOpts where
   dirCustomizationsL = lens xdirCustomizations (\x y -> x { xdirCustomizations = y })
 
@@ -279,10 +284,11 @@ data ProtoOpts = ProtoOpts
    -- ^ when true, an existing prototype directory will be destroyed and re-installed
    , x1OutputLanguage :: !(Maybe Lang)
    , x1fSpecGenOpts :: !FSpecGenOpts
-   , xskipComposer :: !Bool
    , xdirPrototype :: !(Maybe FilePath)
-   , xdirCustomizations :: ![FilePath]
+   , xdirCustomizations :: !(Maybe [FilePath])
    , xzwolleVersion :: !FilePath
+   , xgenerateFrontend :: !Bool
+   , xgenerateBackend :: !Bool
   } deriving Show
 
 -- | Options for @ampersand documentation@.